third_party/blink/web_tests/virtual/saa-beyond-cookies-disabled/external/wpt/storage-access-api/storage-access-beyond-cookies.getDirectory.tentative.sub.https.window-expected.txt

This is a testharness.js-based test.
[FAIL] Verify StorageAccessAPIBeyondCookies for Origin Private File System
  assert_equals: Storage Access API should be accessible and return first-party data expected "HasAccess for getDirectory" but got "Unable to load handle in cross-site context for all"
Harness: the test ran to completion.
